backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pdfplumber==0.11.4
PyMuPDF==1.25.3
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
google-genai>=1.0.0
slowapi==0.1.9
pillow==11.1.0
